requirements.txt
Django==4.2.4
django-cors-headers==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
psycopg2-binary
pycparser==2.21

python-dotenv~=1.0.0